pyproject.toml
[tool.poetry]
name = "repl_python3_PyraBot"
version = "0.1.0"
description = ""
authors = ["ML10wastaken <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.7.4"
async-timeout = "^3.0.1"
googletrans = "^3.0.0"
qrcode = "^7.3"
Flask = "^2.0.1"
youtube_dl = "^2021.6.6"
lyricsgenius = "^3.0.1"
replit = "^3.2.4"
bs4 = "^0.0.1"
discord-ext-wizards = "^0.3.0"
discord-pretty-help = "^1.3.2"
owotext = "^1.1.1"
"dislash.py" = "^1.4.9"
jishaku = "^2.2.0"
psutil = "^5.8.0"
emoji = "^1.4.2"
"discord.py" = "^1.7.3"
PyNaCl = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
